requirements.txt
asgiref==3.2.10
Django==3.0.7
django-cors-headers==3.2.1
django-environ==0.4.5
django-filter==2.2.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
psycopg2==2.8.5
pytz==2020.1
sqlparse==0.3.1